[wip] more cleanup
